pyproject.toml
[project]
name = "snapwright"
version = "0.1.0"
description = "Infrastructure as Code for Behringer Wing mixer snapshots"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.3",
    "deepdiff>=7.0",
]

[project.scripts]
snapwright = "snapwright.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.setuptools.packages.find]
include = ["snapwright*"]
